# servers/assessment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the assessment server of PROMPT.",
        "title": "PROMPT Assessment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/assessment/api",
    "paths": {
...
        "/course_phase/{coursePhaseID}/evaluation/feedback-items": {
            "get": {
                "description": "List feedback items for a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "List feedback items for course phase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/feedbackItemDTO.FeedbackItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a feedback item for the current student.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "Create feedback item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feedback item payload",
                        "name": "feedbackItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feedbackItemDTO.CreateFeedbackItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/evaluation/feedback-items/course-participation/{courseParticipationID}": {
            "get": {
                "description": "List feedback items for a course participation in a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "List feedback items for participant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/feedbackItemDTO.FeedbackItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/evaluation/feedback-items/my-feedback": {
            "get": {
                "description": "List feedback items authored by the current student.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "List my feedback items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/feedbackItemDTO.FeedbackItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/evaluation/feedback-items/tutor/{courseParticipationID}": {
            "get": {
                "description": "List feedback items for a tutor in a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "List feedback items for tutor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tutor course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/feedbackItemDTO.FeedbackItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/evaluation/feedback-items/{feedbackItemID}": {
            "put": {
                "description": "Update a feedback item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback_items"
                ],
                "summary": "Update feedback item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Feedback item ID",
                        "name": "feedbackItemID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Feedback item payload",
                        "name": "feedbackItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feedbackItemDTO.UpdateFeedbackItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a feedback item by ID.",
                "tags": [
                    "feedback_items"
                ],
                "summary": "Delete feedback item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Feedback item ID",
                        "name": "feedbackItemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "assessmentType.AssessmentType": {
            "type": "string",
            "enum": [
                "self",
                "peer",
                "tutor",
                "assessment"
            ],
            "x-enum-varnames": [
                "Self",
                "Peer",
                "Tutor",
                "Assessment"
            ]
        },
...
        "db.FeedbackType": {
            "type": "string",
            "enum": [
                "positive",
                "negative"
            ],
            "x-enum-varnames": [
                "FeedbackTypePositive",
                "FeedbackTypeNegative"
            ]
        },
...
        "feedbackItemDTO.CreateFeedbackItemRequest": {
            "type": "object",
            "required": [
                "authorCourseParticipationID",
                "courseParticipationID",
                "coursePhaseID",
                "feedbackType",
                "type"
            ],
            "properties": {
                "authorCourseParticipationID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "feedbackText": {
                    "type": "string"
                },
                "feedbackType": {
                    "enum": [
                        "positive",
                        "negative"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/db.FeedbackType"
                        }
                    ]
                },
                "type": {
                    "enum": [
                        "self",
                        "peer",
                        "tutor",
                        "assessment"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/assessmentType.AssessmentType"
                        }
                    ]
                }
            }
        },
        "feedbackItemDTO.FeedbackItem": {
            "type": "object",
            "properties": {
                "authorCourseParticipationID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "createdAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "feedbackText": {
                    "type": "string"
                },
                "feedbackType": {
                    "$ref": "#/definitions/db.FeedbackType"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/assessmentType.AssessmentType"
                }
            }
        },
        "feedbackItemDTO.UpdateFeedbackItemRequest": {
            "type": "object",
            "required": [
                "authorCourseParticipationID",
                "courseParticipationID",
                "coursePhaseID",
                "feedbackText",
                "feedbackType",
                "id",
                "type"
            ],
            "properties": {
                "authorCourseParticipationID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "feedbackText": {
                    "type": "string"
                },
                "feedbackType": {
                    "enum": [
                        "positive",
                        "negative"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/db.FeedbackType"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "self",
                        "peer",
                        "tutor",
                        "assessment"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/assessmentType.AssessmentType"
                        }
                    ]
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}